+++ jdevelops-dals/jdevelops-dals-quartz/src/main/java/cn/tannn/jdevelops/quartz/QuartzConfiguration.java
@@ -7,7 +7,10 @@ import cn.tannn.jdevelops.quartz.service.ScheduleService;
 import cn.tannn.jdevelops.quartz.service.ScheduleServiceImpl;
 import org.quartz.Scheduler;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Configurable;
+import org.springframework.boot.autoconfigure.AutoConfiguration;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnWebApplication;
+import org.springframework.boot.context.properties.EnableConfigurationProperties;
 import org.springframework.context.annotation.Bean;
 
 import javax.persistence.EntityManager;
@@ -21,93 +24,13 @@ import javax.persistence.EntityManager;
 public class QuartzConfiguration {
 
     @Bean
-    public QuartzConfig jdbcTemplateConfig() {
+    public QuartzConfig quartzConfig() {
         return new QuartzConfig();
     }
-
-    @Bean
-    public QrtzCronTriggersDao qrtzCronTriggersDao(@Autowired EntityManager entityManager) {
-        return new QrtzCronTriggersDaoImpl(entityManager);
-    }
-
-    @Bean
-    public QrtzJobDetailsDao qrtzJobDetailsDao(@Autowired EntityManager entityManager) {
-        return new QrtzJobDetailsDaoImpl(entityManager);
-    }
-
-
-    @Bean
-    public QrtzTriggersDao qrtzTriggersDao(@Autowired EntityManager entityManager) {
-        return new QrtzTriggersDaoImpl(entityManager);
-    }
-
 
     @Bean
     public AutoRegisterJob autoRegisterJob() {
         return new AutoRegisterJob();
     }
-
-    @Bean
-    public ScheduleService scheduleService(Scheduler scheduler, QrtzJobDetailsDao jobDetailsDao) {
-        return new ScheduleServiceImpl(scheduler,jobDetailsDao);
-    }
-
-
-    @Bean
-    public QrtzBlobTriggersEntity qrtzBlobTriggersEntity() {
-        return new QrtzBlobTriggersEntity();
-    }
-
-    @Bean
-    public QrtzCalendarsEntity qrtzCalendarsEntity() {
-        return new QrtzCalendarsEntity();
-    }
-
-    @Bean
-    public QrtzCronTriggersEntity qrtzCronTriggersEntity() {
-        return new QrtzCronTriggersEntity();
-    }
-
-    @Bean
-    public QrtzFiredTriggersEntity qrtzFiredTriggersEntity() {
-        return new QrtzFiredTriggersEntity();
-    }
-
-    @Bean
-    public QrtzJobDetailsEntity qrtzJobDetailsEntity() {
-        return new QrtzJobDetailsEntity();
-    }
-
-    @Bean
-    public QrtzLocksEntity qrtzLocksEntity() {
-        return new QrtzLocksEntity();
-    }
-
-    @Bean
-    public QrtzPausedTriggerGrpsEntity qrtzPausedTriggerGrpsEntity() {
-        return new QrtzPausedTriggerGrpsEntity();
-    }
-
-    @Bean
-    public QrtzSchedulerStateEntity qrtzSchedulerStateEntity() {
-        return new QrtzSchedulerStateEntity();
-    }
-
-    @Bean
-    public QrtzSimpleTriggersEntity qrtzSimpleTriggersEntity() {
-        return new QrtzSimpleTriggersEntity();
-    }
-
-    @Bean
-    public QrtzSimpropTriggersEntity qrtzSimpropTriggersEntity() {
-        return new QrtzSimpropTriggersEntity();
-    }
-
-    @Bean
-    public QrtzTriggersEntity qrtzTriggersEntity() {
-        return new QrtzTriggersEntity();
-    }
-
-
 
 }
