+++ jdevelops-dals/jdevelops-dals-quartz/src/main/java/cn/tannn/jdevelops/quartz/config/QuartzConfig.java
@@ -1,6 +1,7 @@
 package cn.tannn.jdevelops.quartz.config;
 
 import org.springframework.boot.context.properties.ConfigurationProperties;
+import org.springframework.context.annotation.Configuration;
 
 /**
  * 配置文件
