+++ jdevelops-frameworks/jdevelops-quartz-quick-starter/pom.xml
@@ -5,23 +5,30 @@
     <modelVersion>4.0.0</modelVersion>
     <parent>
         <groupId>cn.tannn.jdevelops</groupId>
-        <artifactId>jdevelops-dals</artifactId>
+        <artifactId>jdevelops-frameworks</artifactId>
         <version>0.0.1</version>
     </parent>
 
-    <artifactId>jdevelops-dals-quartz</artifactId>
+    <artifactId>jdevelops-quartz-quick-starter</artifactId>
     <packaging>jar</packaging>
-    <name>jdevelops :: dals :: quartz</name>
+    <name>jdevelops :: quartz :: quick :: starter</name>
     <version>0.0.1-SNAPSHOT</version>
     <description>
-        quartz 封装
+        quartz  quick starter
     </description>
+
 
     <dependencies>
         <dependency>
             <groupId>org.springframework.boot</groupId>
-            <artifactId>spring-boot-starter-quartz</artifactId>
+            <artifactId>spring-boot-starter-web</artifactId>
             <version>${spring.boot.version}</version>
+            <optional>true</optional>
+        </dependency>
+        <dependency>
+            <groupId>cn.tannn.jdevelops</groupId>
+            <artifactId>jdevelops-dals-quartz</artifactId>
+            <version>0.0.1-SNAPSHOT</version>
         </dependency>
         <dependency>
             <groupId>io.swagger.core.v3</groupId>
@@ -29,21 +36,17 @@
             <optional>true</optional>
         </dependency>
         <dependency>
-            <groupId>org.springframework.boot</groupId>
-            <artifactId>spring-boot-starter-data-jpa</artifactId>
-            <version>${spring.boot.version}</version>
+            <groupId>cn.tannn.jdevelops</groupId>
+            <artifactId>jdevelops-dals-jpa</artifactId>
+            <version>0.0.2</version>
             <optional>true</optional>
         </dependency>
         <dependency>
-            <groupId>org.springframework.boot</groupId>
-            <artifactId>spring-boot-autoconfigure</artifactId>
+            <groupId>cn.tannn.jdevelops</groupId>
+            <artifactId>jdevelops-apis-result</artifactId>
+            <version>0.0.2</version>
             <optional>true</optional>
-        </dependency>
-        <dependency>
-            <groupId>org.reflections</groupId>
-            <artifactId>reflections</artifactId>
-            <version>0.10.2</version>
-            <scope>compile</scope>
         </dependency>
     </dependencies>
+
 </project>
