+++ jdevelops-frameworks/jdevelops-quartz-quick-starter/src/main/resources/META-INF/spring/org.springframework.boot.autoconfigure.AutoConfiguration.imports
@@ -1,0 +1,1 @@
+cn.tannn.jdevelops.quick.quratz.QuickQuartzConfiguration
