+++ jdevelops-frameworks/pom.xml
@@ -59,6 +59,7 @@
     <modules>
         <module>jdevelops-spring-boot-starter</module>
         <module>jdevelops-file-starter</module>
+        <module>jdevelops-quartz-quick-starter</module>
     </modules>
